packages/shardon_core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "shardon-core"
version = "0.1.0"
description = "Shared domain logic for the Shardon monorepo."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/shardon_core"]

